+++ Assets/_Scripts/GameManager.cs
@@ -1,46 +1,90 @@
+using _Scripts.Ball;
 using UnityEngine;
 
-public class GameManager : MonoBehaviour
+namespace _Scripts
 {
-    [SerializeField] private GameObject ballPrefab;
-
-    [SerializeField] private int maxPlayerLive = 3;
-    private int _currentPlayerLive;
-
-    // UI
-    [SerializeField] private Canvas scoreboard;
-    [SerializeField] private Canvas victoryScreen;
-
-    // Start is called before the first frame update
-    void Awake()
+    public class GameManager : MonoBehaviour
     {
-        _currentPlayerLive = maxPlayerLive;
-    }
+        [SerializeField] private GameObject ballPrefab;
+        private static GameManager instance;
 
-    // Update is called once per frame
-    void Update()
-    {
-        if (_currentPlayerLive == 0)
+        [SerializeField] private int maxPlayerLive = 3;
+        public static int _currentPlayerLive;
+        public static int highscore;
+        public static float time;
+        
+        public static bool running = false;
+
+        // UI
+        [SerializeField] private Canvas scoreboard;
+        [SerializeField] private Canvas victoryScreen;
+
+        public static GameManager Instance
         {
-            Loose();
+            get
+            {
+                return instance;
+            }
         }
-    }
 
-    private void Loose()
-    {
-        scoreboard.enabled = false;
-        victoryScreen.enabled = true;
-    }
+        void Awake()
+        {
+            time = 0;
+            _currentPlayerLive = maxPlayerLive;
+        }
 
-    private void SpawnBall()
-    {
-        Debug.Log("Respawned Ball");
-        Instantiate(ballPrefab, new Vector3(0,0.25f,0), Quaternion.identity);
-    }
+        // Update is called once per frame
+        void Update()
+        {
+            if (Input.GetKey(KeyCode.Space) && !running)
+            {
+                Debug.Log("Space");
+                StartGame();
+            }
 
-    public void TakeDamage()
-    {
-        _currentPlayerLive -= 1;
-        if (_currentPlayerLive > 0) SpawnBall();
+            if (running)
+            {
+                time += Time.deltaTime;
+            }
+            
+            if (_currentPlayerLive == 0)
+            {
+                Loose();
+            }
+        }
+
+        private void StartGame()
+        {
+            GameObject.FindWithTag("Ball").GetComponent<BallController>().Launch();
+            StartTimer();
+        }
+
+        private void Loose()
+        {
+            scoreboard.enabled = false;
+            victoryScreen.enabled = true;
+            StopTimer();
+            Time.timeScale = 0;
+        }
+
+        public void TakeDamage()
+        {
+            _currentPlayerLive -= 1;
+            if (_currentPlayerLive > 0)
+            {
+                // Spawn new ball
+                Instantiate(ballPrefab, new Vector3(0,0.25f,-4.1f), Quaternion.identity);
+            }
+        }
+        
+        private void StartTimer()
+        {
+            running = true;
+        }
+
+        public void StopTimer()
+        {
+            running = false;
+        }
     }
 }