+++ Assets/_Scripts/PowerUps/PowerUpType.cs
@@ -1,0 +1,8 @@
+﻿namespace _Scripts.PowerUps
+{
+    public enum PowerUpType
+    {
+        IncreasePaddleSize,
+        IncreasePaddleSpeed
+    }
+}