+++ Changelog.md
@@ -2,12 +2,45 @@
 
 ## [Unreleased]
 
-## [0.0.1] - 2022-09-03
+## [0.6.0] - 2022-10-17
+
+### Added
+- Correct behaviour of powerUps when being spawned
+
+
+## [0.5.0] - 2022-10-15
+- Code Refactoring
+
+### Added
+- Correct behaviour of paddle when reaching edge of playfield
+- Correct behaviour of ball when reaching top of playfield and paddle
+- Game starts when pressing Space
+- Ball sticks to paddle until game starts
+- Implemented a Main Menu
+- random spawning of powerUps when blocks are destroyed
+
+
+## [0.4.0] - 2022-10-13
+### Added
+- Correct behaviour of blocks when getting hit
+
+
+## [0.3.0] - 2022-10-12
+### Added
+- Blocks
+
+
+## [0.2.0] - 2022-10-10
+### Added
+- Score
+- PlayerLife is now visible
+- Correct behaviour of Timer
+
+
+## [0.1.0] - 2022-09-03
 ### Added
 - PlayerLife
 - Win/Loose UI
 - Timer
-
-### Changed
 
 ### Removed